requirements/production.txt
alembic==1.8.0
argon2-cffi==21.3.0
asyncpg==0.25.0
dependency-injector==4.39.1
fastapi-mail==1.0.9
fastapi==0.78.0
passlib==1.7.4
psycopg2-binary==2.9.3
PyJWT==2.4.0
PyYAML==6.0
sqlmodel==0.0.6
uvicorn==0.18.1
pytest-asyncio==0.18.3